requirements.txt
cycler==0.10.0
descartes==1.0.2
Django==1.8.1
django-admin-bootstrapped==2.5.7
django-bootstrap-toolkit==2.15.0
django-bootstrap3==7.1.0
django-chartit==0.2.8
django-crispy-forms==1.6.1
django-geojson==2.9.1
django-grappelli==2.9.1
django-leaflet==0.19.0
django-markdown==0.8.4
django-registration-redux==1.4
djangorestframework==3.5.3
easy-thumbnails==2.3
geojson==1.3.3
Markdown==2.6.7
matplotlib==1.5.3
numpy==1.11.2
pandas==0.19.0
Pillow==3.4.2
psycopg2==2.6.2
pyparsing==2.1.10
pyproj==1.9.5.1
PySAL==1.12.0
pyshp==1.2.10
python-dateutil==2.5.3
pytz==2016.7
reportlab==3.3.0
scipy==0.18.1
Shapely==1.5.17
simplejson==3.10.0
six==1.10.0
